+++ R/extras.R
@@ -287,11 +287,7 @@ cheapr_var <- function(x, na.rm = TRUE){
 #' @rdname extras
 #' @export
 cheapr_rev <- function(x){
-  if (cpp_is_simple_vec(x)){
-    .Call(`_cheapr_cpp_sset`, x, length(x):0, TRUE)
-  } else {
-    sset(x, vector_length(x):0)
-  }
+  .Call(`_cheapr_cpp_sset`, x, vector_length(x):0, TRUE)
 }
 cheapr_sd <- function(x, na.rm = TRUE){
   sqrt(cheapr_var(x, na.rm = na.rm))