+++ R/factors.R
@@ -401,7 +401,7 @@ levels_count <- function(x){
   check_is_factor(x)
   names <- attr(x, "levels")
   out <- tabulate(x, nbins = length(names))
-  fast_df(name = names, count = out, prop = out / sum(out))
+  new_df(name = names, count = out, prop = out / sum(out))
 }
 # Generic factor conversion to data representation
 factor_as_type <- function(x, type){